requirements.txt
# Collecte de données
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
pandas==2.2.2
arxiv==2.0.0
scholarly==1.7.11
python-dotenv==1.0.1
tqdm==4.66.2

# Analyse de données
transformers==4.43.2
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
wordcloud==1.9.3
nltk==3.8.1
spacy==3.7.4
openai==1.25.1

# Diffusion et Telegram
python-telegram-bot==20.7
schedule==1.2.1

# Génération de podcast
pydub==0.25.1

# Dépendances spécifiques
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl